changed spi_fifo_read, and get_fifo_data to 8bit param and dma

+++ Src/xensiv_bgt60trxx.c
@@ -288,7 +288,7 @@ int32_t xensiv_bgt60trxx_set_fifo_limit(const xensiv_bgt60trxx_t* dev, uint32_t 
 }
 
 
-int32_t xensiv_bgt60trxx_get_fifo_data(xensiv_bgt60trxx_t* dev, uint16_t* data,
+int32_t xensiv_bgt60trxx_get_fifo_data(xensiv_bgt60trxx_t* dev, uint8_t* data,
                                        uint32_t num_samples)
 {
     xensiv_bgt60trxx_platform_assert(dev != NULL);
